fix integer type
svn path=/trunk/yarp2/; revision=8884

+++ src/carriers/xmlrpc_carrier/xmlrpc/XmlRpcClient.cpp
@@ -254,7 +254,7 @@ XmlRpcClient::generateHeader(std::string const& body)
   header += buff;
   header += "Content-Type: text/xml\r\nContent-length: ";
 
-  sprintf(buff,"%d\r\n\r\n", body.size());
+  sprintf(buff,"%d\r\n\r\n", (int)body.size());
 
   return header + buff;
 }
